% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefAll2.R
\name{coefAll2,ANY-method}
\alias{coefAll2,ANY-method}
\title{Method for extracting coefficients from GAM-like models}
\usage{
\S4method{coefAll2}{ANY}(obj, deviance = FALSE)
}
\arguments{
\item{obj}{A model object, typically from a GAM-like fitting procedure.}

\item{deviance}{A Boolean flag that when TRUE calculates the deviance of the model.}
}
\value{
A list containing model coefficients and optionally deviance.
}
\description{
Method for extracting coefficients from GAM-like models
}
