% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefAll2.R
\docType{methods}
\name{coefAll2}
\alias{coefAll2}
\title{Extract Coefficients}
\usage{
coefAll2(obj, deviance = FALSE)
}
\arguments{
\item{obj}{An object.}

\item{deviance}{Logical value indicating whether to compute deviance.}
}
\value{
A list of coefficients.
}
\description{
Internal method to extract coefficients from an object.
}
\examples{
print("hello")

}
