% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_families.R
\docType{data}
\name{copula_families}
\alias{copula_families}
\alias{.copula_families}
\title{Define Copula Families and Parameters}
\format{
An object of class \code{list} of length 3.
}
\usage{
.copula_families
}
\description{
Internal list of defined copula families and their corresponding parameters.
}
\keyword{datasets}
