% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_pbox.R
\name{final_pbox}
\alias{final_pbox}
\title{Fit the \code{mvdc} object.}
\usage{
final_pbox(results_df, allDitrs, data)
}
\arguments{
\item{results_df}{The data.table with the correspondent AIC and the parameter estimates of the evaluated copulas and families resulting from \code{fit_copula_pbox}.}

\item{allDitrs}{The list of the fitted distributions for each variable resulting from \code{fit_dist_pbox}.}

\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}
}
\value{
An object of class \code{mvdc}.
}
\description{
Combine all the results from \code{fit_copula_pbox} and \code{fit_dist_pbox} to build a Multivariate Distributions from Copula.
}
\examples{
SEAex<-fread("./data/SEAex.csv")
copulaFits<- fit_copula_pbox(data=SEAex,copula_families)
distFits<- fit_dist_pbox(data=SEAex)

final_pbox(copulaFits,distFits$allDitrs,SEAex)

}
