% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_copula.R
\docType{methods}
\name{fit_copula}
\alias{fit_copula}
\title{Copula Fit}
\usage{
.fit_copula(copula, family, dim, u)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}

\item{.copula_families}{List of copula types and their corresponding families. Currently supported families are "clayton", "frank", "amh", "gumbel", and "joe" for Archimedean Copula; "galambos", "gumbel", and "huslerReiss" for Extreme-Value copula; "normal" and "t" for Elliptical copula.}
}
\value{
A \code{data.table} with the corresponding AIC and the parameter estimates of the evaluated copulas and families.
}
\description{
Internal method to automatically find the best Copula given a data.frame. Wrapper around the function \code{fitCopula}.
}
\examples{
SEAex <- fread("./data/SEAex.csv")
# Define the copula families and their corresponding parameters
.copula_families <- list(
  archmCopula = c("clayton", "frank", "gumbel", "joe"),
  evCopula = c("galambos", "gumbel", "huslerReiss"),
  ellipCopula = c("normal")
)
distFits <- fit_copula(data = SEAex, .copula_families)
distFits

}
