% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dist_pbox.R
\name{fit_dist_pbox}
\alias{fit_dist_pbox}
\title{Marginal Distributions Fit}
\usage{
fit_dist_pbox(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}

\item{...}{Other arguments to be passed to \code{fitDist}.}
}
\value{
a list containing 2 elements:

\code{allDitrs}: List of the fitted distributions for each variable.

\code{distTable}: \code{data.frame} with numeric value with the corresponding Akaike's Information Criterion \bold{AIC} for each distribution tested.
}
\description{
Method to automatically find the best marginal distribution of each variable in a data.frame.
}
\examples{
SEAex<-fread("./data/SEAex.csv")
distFits<- fit_dist_pbox(data=SEAex)
distFits$allDitrs
distFits$distTable

}
