% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_stats.R
\name{fun_stats,numeric-method}
\alias{fun_stats,numeric-method}
\title{Summary statistics method for numeric vectors}
\usage{
\S4method{fun_stats}{numeric}(x)
}
\arguments{
\item{x}{Numeric vector for which summary statistics are computed.}
}
\value{
A list with components min, max, mean, and median.
}
\description{
This method is a specific implementation of the 'fun_stats' function for numeric vectors.
It efficiently calculates and returns summary statistics including the minimum, maximum,
mean, and median, excluding NA values.
}
