% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_stats.R
\docType{methods}
\name{fun_stats}
\alias{fun_stats}
\title{Summary Statistics}
\usage{
fun_stats(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list containing the minimum, maximum, mean, and median of the input vector.
}
\description{
Internal method to compute summary statistics of a numeric vector.
}
\examples{
x <- c(1, 2, 3, 4, 5)
fun_stats(x)


}
