% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_scenario.R
\name{gen_scenario}
\alias{gen_scenario}
\alias{gen_scenario,ANY-method}
\title{Generate Scenarios}
\usage{
gen_scenario(params = "list")

\S4method{gen_scenario}{ANY}(params = "list")
}
\arguments{
\item{params}{List of parameters where each parameter can vary across scenarios.}
}
\value{
Nested list of scenarios.
}
\description{
Internal method to Generate scenarios based on parameter list variations.
}
\examples{
some_distr<-list(A=list(mu = 31.07, sigma = 0.28),
B=list(mu = c(34.4,31.4,25.6), sigma = 0.98, nu = 1.7),# note mu!
C=list(mu = 31.4, sigma = 0.34),
D=list(mu = 25.6, sigma = 0.24))
gen_scenario(some_distr)
}
