% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pbox.R
\docType{methods}
\name{make_pbox}
\alias{make_pbox}
\title{Make a Pbox}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}

\item{cop}{An object of class \code{mvdc}.}
}
\value{
An object of class \code{pbox} with 2 slots:

\describe{
  \item{\code{data}}{The original data coerced to a data.table.}
  \item{\code{cop}}{The copula object of class \code{mvdc}.}
}
}
\description{
Auxiliary method to create a pbox object from a dataframe and a custom copula object.
}
\examples{
library(copula)
data("SEAex")

cop<-normalCopula(param = 0.5,dim = 4)
distList<-c("RG" , "SN1", "RG","RG" )
allDistrs<-list(list(mu = 31.07, sigma = 0.28),
               list(mu = 34.4, sigma = 0.98, nu = 1.7),
               list(mu = 31.4, sigma = 0.34),
               list(mu = 25.6, sigma = 0.24))
copSEA <- mvdc(cop, distList,allDistrs)
pbx<-make_pbox(data=SEAex,cop=copSEA)
class(pbx)

}
