% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_maker.R
\name{match_maker,ANY-method}
\alias{match_maker,ANY-method}
\title{Method for match_maker}
\usage{
\S4method{match_maker}{ANY}(varSet, matches, data)
}
\arguments{
\item{varSet}{A data frame or list describing the variable set.}

\item{matches}{A data frame describing the matches with variable names and corresponding values or operators.}

\item{data}{A data frame representing the data to be queried.}
}
\value{
A modified version of `varSet` that integrates conditions or values from `matches`.
}
\description{
This method implements the `match_maker` function for handling specific types
of 'varSet', 'matches', and 'data'. It modifies the 'varSet' based on 'matches' which
can contain variable names and values to be matched or operations to be performed.
It supports operations and direct value assignment.
}
\seealso{
\code{\link{match_maker}} for the generic function and additional details.
}
