% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_maker.R
\docType{methods}
\name{match_maker}
\alias{match_maker}
\title{Generate Query Vector}
\usage{
match_maker(varSet, matches, data)
}
\arguments{
\item{varSet}{A string describing the variable set.}

\item{matches}{A description of matches.}

\item{data}{A description of the data.}
}
\value{
A table with the value to be queried of each column/marginal distribution of the data.
}
\description{
Internal function used to generate the query vector to explore the probabilistic space.
}
