% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_dev.R
\name{param_dev}
\alias{param_dev}
\alias{param_dev,ANY-method}
\title{Compute Parameter Deviations}
\usage{
param_dev(param = "numeric", sigma = 0.05, range = seq(-3, 3, 1))

\S4method{param_dev}{ANY}(param = "numeric", sigma = 0.05, range = seq(-3, 3, 1))
}
\arguments{
\item{param}{Numeric vector of parameters.}

\item{sigma}{Numeric value representing standard deviation (default is 0.05).}

\item{range}{Numeric vector specifying range of deviations (default is seq(-3, 3, 1)).}
}
\value{
Numeric vector of parameters adjusted by the specified deviations.
}
\description{
Internal method to calculate ± 1, 2, 3 standard deviations for given parameters.
}
\examples{
param_dev(31)
}
