% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbox.R
\docType{class}
\name{pbox-class}
\alias{pbox-class}
\title{Class "pbox": Main S4 class of the library \bold{pbox}.}
\description{
"pbox" is a class representing the probabilistic space which combines data, copula and margins.
}
\section{Slots}{

\describe{
\item{\code{data}}{The original data coerced to a \code{data.table}.}

\item{\code{copula}}{The copula object of class \code{mvdc}.}

\item{\code{fit}}{The results of the automated selection for both the marginal distribution and the copula.}
}}

