% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbate_params.R
\docType{methods}
\name{perturb_params}
\alias{perturb_params}
\title{Perturb Parameters}
\usage{
perturbate_params(paramMargins)
}
\arguments{
\item{paramMargins}{A list of parameter values for each distribution in the copula.}
}
\value{
A list of perturbed parameter values.
}
\description{
Method to perturb parameters of copula distributions.
}
\examples{
paramMargins <- list(list(0.2, 0.3), list(0.4, 0.5))
perturb_params(paramMargins)

}
