% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbate_params.R
\name{perturbate_params,ANY-method}
\alias{perturbate_params,ANY-method}
\title{Perturb Parameters Method}
\usage{
\S4method{perturbate_params}{ANY}(paramMargins)
}
\arguments{
\item{paramMargins}{A list containing lists of parameter values for each distribution in the copula.}
}
\value{
A list of lists containing perturbed parameter values.
}
\description{
This method implements the generic `perturbate_params` function specifically
for lists of copula distribution parameters. It applies a random perturbation
to each parameter based on a normal distribution centered at zero with a
standard deviation of 0.05.
}
\seealso{
\code{\link{perturbate_params}} for the generic function definition.
}
