% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturbate_params.R
\name{perturbate_params}
\alias{perturbate_params}
\title{Perturb Parameters}
\usage{
perturbate_params(paramMargins)
}
\arguments{
\item{paramMargins}{A list containing lists of parameter values for each distribution in the copula.}
}
\value{
A list of lists containing perturbed parameter values.
}
\description{
This function defines a generic function to perturbate parameter values for each distribution
within a copula, using random perturbations to simulate variability or uncertainty.
}
\examples{
paramMargins <- list(list(0.2, 0.3), list(0.4, 0.5))
perturbed <- perturbate_params(paramMargins)
print(perturbed)
}
