% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCI.R
\name{probCI,numeric-method}
\alias{probCI,numeric-method}
\title{Method to calculate confidence intervals for a vector of probabilities}
\usage{
\S4method{probCI}{numeric}(probabilities, alpha = 0.05)
}
\arguments{
\item{probabilities}{A numeric vector of probabilities.}

\item{alpha}{A numeric value specifying the significance level for the confidence intervals; defaults to 0.05.}
}
\value{
A numeric vector containing the lower and upper quantile bounds for each probability in the input vector.
}
\description{
This method calculates the lower and upper bounds of the confidence interval for each element in the
input vector of probabilities using the given alpha level.
}
