% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probCI.R
\docType{methods}
\name{probCI}
\alias{probCI}
\title{Probability Confidence Interval}
\usage{
probCI(probabilities, alpha = 0.05)
}
\arguments{
\item{probabilities}{A numeric vector of probabilities.}

\item{alpha}{The significance level for the confidence interval. Default is 0.05.}
}
\value{
A numeric vector containing the lower and upper bounds of the confidence interval.
}
\description{
Method to calculate the confidence interval around probabilities.
}
\examples{
probabilities <- c(0.1, 0.2, 0.3, 0.4, 0.5)
probCI(probabilities)
probCI(probabilities, alpha = 0.1)

}
