% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_parser.R
\name{q_parser,ANY-method}
\alias{q_parser,ANY-method}
\title{Method for Parsing Queries}
\usage{
\S4method{q_parser}{ANY}(query)
}
\arguments{
\item{query}{A string representing the query.}
}
\value{
A data table with the parsed elements of the query.
}
\description{
Implements the `q_parser` function specifically for string input. It uses a regular
expression to split the query into its components, converting numeric strings to numeric
values where applicable, and structuring the result as a data table for easy manipulation.
}
\seealso{
\code{\link{q_parser}} for the generic function definition.
}
