% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_parser.R
\docType{methods}
\name{q_parser}
\alias{q_parser}
\title{Parse query to explore a pbox object}
\usage{
q_parser(query)
}
\arguments{
\item{query}{string}
}
\value{
A table with the parsed elements of the query.
}
\description{
Internal function used to Parse query to explore a pbox object.
}
