% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpbox.R
\docType{methods}
\name{qpbox}
\alias{qpbox}
\title{Query the probabilistic space of a pbox object.}
\usage{
qpbox(
  x,
  marginal = "character",
  conditional = "character",
  lower.tail = TRUE,
  fixed = FALSE,
  CI = FALSE,
  iter = 1000
)
}
\arguments{
\item{x}{object of class \bold{pbox} from which to query the probabilistic space.}

\item{marginal}{character; string used to query the marginal and joint distribution of the variable. Must specify the variable and the value in the following format 'Var:Val'}

\item{conditional}{character; string used to query the marginal and conditional distribution of the variable. Must specify the variable and the value in the following format 'Var:Val'}

\item{lower.tail}{logical; if TRUE (default), probabilities are calculated for the area to the right of the specified value.}

\item{fixed}{logical; if TRUE (default), probabilities are \eqn{P(X \leq x | Y = y)}}

\item{CI}{logical; if TRUE, calculate bootstrap confidence intervals. See notes for details.}

\item{iter}{number of replications for the confidence interval calculation. 1000 by default.}
}
\value{
Estimated probability.
}
\description{
Combine all the results from \code{fit_copula_pbox} and \code{fit_dist_pbox} to build a Multivariate Distributions from Copula.
}
