% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_pbox.R
\name{set_pbox}
\alias{set_pbox}
\title{Create a probability box from data.}
\usage{
set_pbox(data, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}

\item{...}{Other arguments to be passed to \code{fitDist}.}
}
\value{
an object of class \code{pbox} with 3 slots:

\code{data}: The original data coerced to a data.table.

\code{copula}: The copula object of class \code{mvdc}.

\code{fit}: The results of the automated selection for both the marginal distribution and the copula.
}
\description{
Method to automatically find the best marginal distribution and copula for a give dataset.
}
\examples{
SEAex<-fread("./data/SEAex.csv")
pbx<- set_pbox(data=SEAex)
pbx
class(pbx)

}
