% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pbox.R
\docType{methods}
\name{make_pbox}
\alias{make_pbox}
\title{Make a Pbox}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}

\item{copula}{An object of class \code{mvdc}.}
}
\value{
An object of class \code{pbox} with 2 slots:

\describe{
  \item{\code{data}}{The original data coerced to a data.table.}
  \item{\code{copula}}{The copula object of class \code{mvdc}.}
}
}
\description{
Auxiliary method to create a pbox object from a dataframe and a custom copula object.
}
\examples{
data("SEAex")
pbx <- make_pbox(data = SEAex, copula = some_mvdc_object)
pbx
class(pbx)

}
