% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pbox.R
\name{make_pbox}
\alias{make_pbox}
\title{Make a pbox from a data.frame and a copula object.}
\usage{
make_pbox(data, copula)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}
}
\value{
an object of class \code{pbox} with 2 slots:

\code{data}: The original data coerced to a data.table.

\code{copula}: The copula object of class \code{mvdc}.
}
\description{
Method to allow users to query a copula object created by the,
}
\examples{
SEAex<-fread("./data/SEAex.csv")

pbx<- set_pbox(data=SEAex)
pbx
class(pbx)

}
