% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pbox.R
\docType{methods}
\name{make_pbox}
\alias{make_pbox}
\title{Make a pbox from a data.frame and a copula object.}
\usage{
make_pbox(data, copula)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table} (the data will be coerced to a \code{data.table} internally).}

\item{copula}{an object of class mvdc.}
}
\value{
an object of class \code{pbox} with 2 slots:

\code{data}: The original data coerced to a data.table.

\code{copula}: The copula object of class \code{mvdc}.
}
\description{
Ausiliary method to create a pbox object from a dataframe and a custom copula object.
}
\examples{
data("SEAex")

pbx<- set_pbox(data=SEAex)
pbx
class(pbx)

}
