% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_maker.R
\docType{methods}
\name{match_maker}
\alias{match_maker}
\title{Generate the query vector to explore the probabilistic space.}
\usage{
match_maker(varSet, matches, data)
}
\arguments{
\item{varSet}{string}

\item{matches}{description}

\item{data}{description}
}
\value{
A table with the value to be queried of each column/marginal distribution of the data.
}
\description{
Internal function used to generate the query vector to explore the probabilistic space.
}
