% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_maker.R
\name{match_maker}
\alias{match_maker}
\title{Generate the query vector to explore the probabilistc space.}
\usage{
match_maker(varSet, matches, data)
}
\arguments{
\item{varSet}{string}

\item{matches}{description}

\item{data}{description}
}
\value{
A table with the parsed elements of the query.
}
\description{
Internal function used to Parse query to explore a pbox object.
}
