% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_maker.R
\docType{methods}
\name{match_maker}
\alias{match_maker}
\title{Generate Query Vector}
\usage{
match_maker(varSet, matches, data)
}
\arguments{
\item{varSet}{A data frame or list describing the variable set.}

\item{matches}{A data frame describing the matches with potential additional control parameters.}

\item{data}{A data frame representing the data to be queried.}
}
\value{
A modified version of `varSet` with values updated based on `matches`.
}
\description{
This function defines a generic function for creating a query vector to explore
the probabilistic space based on provided matches and data. It is used internally
to handle different types of inputs efficiently.
}
