% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perProb.R
\docType{methods}
\name{perProb}
\alias{perProb}
\title{Perturbed Probability}
\usage{
perProb(x, vecQuery)
}
\arguments{
\item{x}{An object of class \code{pbox} representing the base copula.}

\item{vecQuery}{A numeric vector representing the query values.}
}
\value{
The probability computed using a perturbed copula.
}
\description{
Internal method to compute the probability using a perturbed copula.
}
\examples{
bpx <- some_mvdc_object
vecQuery <- c(0.5, 0.5)
.perProb(bpx, vecQuery)

}
