% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_parser.R
\docType{methods}
\name{q_parser}
\alias{q_parser}
\title{Parse Query}
\usage{
q_parser(query)
}
\arguments{
\item{query}{A string representing the query.}
}
\value{
A table with the parsed elements of the query.
}
\description{
Internal function used to parse a query to explore a pbox object.
}
\examples{
query <- "some_query_string"
q_parser(query)

}
