% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_parser.R
\name{q_parser}
\alias{q_parser}
\title{Parse Query}
\usage{
q_parser(query)
}
\arguments{
\item{query}{A string representing the query.}
}
\value{
A data table with columns 'Varnames', 'Value', 'Operator', and 'Varnames2',
        where numeric values are converted to numeric type, and unnecessary columns are removed.
}
\description{
This function defines a generic function to parse a string query into structured
data that can be used to explore a pbox object. It extracts components of the query
using regular expression matching.
}
\examples{
query <- "Vietnam:23"
q_parser(query)
}
