% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpbox.R
\docType{methods}
\name{qpbox}
\alias{qpbox}
\title{Query the probabilistic space of a pbox object.}
\usage{
qpbox(
  x,
  marginal = "character",
  conditional = "character",
  lower.tail = TRUE,
  fixed = FALSE
)
}
\arguments{
\item{x}{object of class \bold{pbox} from which to query the probabilistic space.}

\item{marginal}{character; string used to query the marginal and joint distribution of the variable. Must specify the variable and the value in the following format 'Var:Val'}

\item{conditional}{character; string used to query the marginal and conditional distribution of the variable. Must specify the variable and the value in the following format 'Var:Val'}

\item{lower.tail}{logical; if TRUE (default), probabilities are $$}

\item{fixed}{logical; if TRUE (default), probabilities are $$}
}
\value{
An object of class \code{mvdc}.
}
\description{
Combine all the results from \code{fit_copula_pbox} and \code{fit_dist_pbox} to build a Multivariate Distributions from Copula.
}
\examples{
data("SEAex")
pbx<-set_pbox(SEAex)
#Get marginal distribution
pbx["Malaysia:33",]
#Get Joint distribution
pbx["Malaysia:33 & Vietnam:34",]
#Get Joint distribution
pbx["Vietnam:31", "avgRegion:26"]
#Conditional distribtuion Pr(X <= x, Y <= y) / Pr(Y <= y)
pbx["Malaysia:33 & Vietnam:31", "avgRegion:26"]
#Conditional distribtuion Pr(X <= x, Y <= y) / Pr(Y = y)
pbx["Malaysia:33 & Vietnam:31", "avgRegion:26",fixed=TRUE]
# Joint distribution with values set on their respective mean value
pbx["mean:c(Vietnam,Thailand)",lower.tail=T]
# Joint distribution with values set on their respective median value
pbx["median:c(Vietnam, Thailand)",lower.tail=T]
# Joint distribution with xxxx
pbx["Malaysia:33 & mean:c(Vietnam, Thailand)",lower.tail=T]
# Condtional distribtuion distribution with Pr(X <= x, Y <= y) / Pr(Y = y)
pbx["Malaysia:33 & median:c(Vietnam,Thailand)", "mean:c(avgRegion)"]

}
