% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calc.R
\docType{methods}
\name{stats_calc}
\alias{stats_calc}
\title{Calculate Basic Statistics}
\usage{
stats_calc(data, matches, varSet)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.table}.}

\item{matches}{A \code{data.frame} with user query.}

\item{varSet}{A \code{data.frame} with variable names.}
}
\value{
A modified \code{varSet} with the calculated results.
}
\description{
Internal function used to calculate mean and median as part of the query.
}
