% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calc.R
\name{stats_calc}
\alias{stats_calc}
\title{Calculate basic statistics to be used to query the pbox object}
\usage{
stats_calc(data, matches, varSet)
}
\arguments{
\item{varSet}{}
}
\description{
Internal function used to calculate mean and median as part of the query.
}
