% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_calc.R
\name{stats_calc}
\alias{stats_calc}
\alias{stats_calc,ANY-method}
\title{Calculate Basic Statistics}
\usage{
stats_calc(data, matches, varSet)

\S4method{stats_calc}{ANY}(data, matches, varSet)
}
\arguments{
\item{data}{A data frame or data table.}

\item{matches}{A data frame describing the operations to apply.}

\item{varSet}{A data frame to be updated with results.}
}
\value{
Returns a modified version of `varSet` with updated values based on the calculations.
}
\description{
Computes basic statistics such as mean and median for specified variables in a data frame or data table
based on a set of operations specified in the `matches` data frame. This function updates the `varSet`
with the computed results for each variable.

Method implementation for calculating statistics using 'data.table' and 'stats'.
This method allows the computation of mean and median for subsets of data defined
in `matches` and updates `varSet` with these results.
}
